Fix #803: Use relational patterns for VB Select Case
When converting a Visual Basic Select Case statement with a relational
clause (e.g., `Case Is < 1000`), the C# code converter used to output a
variable pattern with a when clause (e.g., `case var @case when @case < 1000:`).
With modern C#, this can be represented more concisely using relational
patterns (e.g., `case < 1000:`).

This commit modifies `MethodBodyExecutableStatementVisitor.cs` to map
Visual Basic relational case clauses to C# `RelationalPatternSyntax`,
`ConstantPatternSyntax` (for equality), and `UnaryPatternSyntax` (for
inequality using `not`).

It also updates existing tests that were affected by this change to
expect the new, cleaner syntax, and adds a test specific to Issue 803.
